requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
pydantic[email]==2.4.2
httpx==0.25.1
asyncpg==0.27.0